separate out DiscoGapicParser, containing static util methods, from DiscoGapicNamer
