+++ src/main/java/com/google/api/codegen/config/DiscoGapicInterfaceConfig.java
@@ -17,6 +17,7 @@ package com.google.api.codegen.config;
 import com.google.api.codegen.CollectionConfigProto;
 import com.google.api.codegen.InterfaceConfigProto;
 import com.google.api.codegen.MethodConfigProto;
+import com.google.api.codegen.discogapic.transformer.DiscoGapicParser;
 import com.google.api.codegen.discovery.Document;
 import com.google.api.codegen.discovery.Method;
 import com.google.api.codegen.transformer.RetryDefinitionsTransformer;
@@ -136,7 +137,7 @@ public abstract class DiscoGapicInterfaceConfig implements InterfaceConfig {
     if (methodConfigs != null) {
       for (MethodConfig methodConfig : methodConfigs) {
         Method method = ((DiscoveryMethodModel) methodConfig.getMethodModel()).getDiscoMethod();
-        String canonicalMethodPath = model.getDiscoGapicNamer().getCanonicalPath(method);
+        String canonicalMethodPath = DiscoGapicParser.getCanonicalPath(method);
         for (SingleResourceNameConfig nameConfig : singleResourceNames) {
           if (nameConfig.getNamePattern().equals(canonicalMethodPath)) {
             methodToSingleResourceNameMap.put(methodConfig, nameConfig);
@@ -150,10 +151,7 @@ public abstract class DiscoGapicInterfaceConfig implements InterfaceConfig {
     String interfaceName =
         interfaceNameOverride != null
             ? interfaceNameOverride
-            : model
-                .getDiscoGapicNamer()
-                .getInterfaceName(interfaceConfigProto.getName())
-                .toUpperCamel();
+            : DiscoGapicParser.getInterfaceName(interfaceConfigProto.getName()).toUpperCamel();
 
     if (model.getDiagCollector().hasErrors()) {
       return null;
