+++ src/main/java/com/google/api/codegen/configgen/transformer/DiscoConfigTransformer.java
@@ -22,7 +22,7 @@ import com.google.api.codegen.configgen.viewmodel.InterfaceView;
 import com.google.api.codegen.configgen.viewmodel.LanguageSettingView;
 import com.google.api.codegen.configgen.viewmodel.LicenseView;
 import com.google.api.codegen.configgen.viewmodel.ResourceNameGenerationView;
-import com.google.api.codegen.discogapic.transformer.DiscoGapicNamer;
+import com.google.api.codegen.discogapic.transformer.DiscoGapicParser;
 import com.google.api.codegen.discovery.Document;
 import com.google.api.codegen.discovery.Method;
 import com.google.api.codegen.util.Name;
@@ -54,7 +54,6 @@ public class DiscoConfigTransformer {
       new MethodTransformer(new DiscoveryMethodTransformer());
 
   public ViewModel generateConfig(DiscoApiModel model, String outputPath) {
-    DiscoGapicNamer namer = model.getDiscoGapicNamer();
     // Map of methods to unique resource names.
     Map<Method, Name> methodToResourceName = new HashMap<>();
 
@@ -64,14 +63,14 @@ public class DiscoConfigTransformer {
     // Maps visited simple resource names to canonical name patterns. Used to check for collisions in simple resource names.
     Map<Name, String> simpleResourceToFirstPatternMap = new HashMap<>();
     for (Method method : model.getDocument().methods()) {
-      String namePattern = model.getDiscoGapicNamer().getCanonicalPath(method);
+      String namePattern = DiscoGapicParser.getCanonicalPath(method);
       methodToNamePattern.put(method, namePattern);
 
-      Name simpleResourceName = namer.getResourceIdentifier(method.flatPath());
+      Name simpleResourceName = DiscoGapicParser.getResourceIdentifier(method.flatPath());
       String collisionPattern = simpleResourceToFirstPatternMap.get(simpleResourceName);
       if (collisionPattern != null && !collisionPattern.equals(namePattern)) {
         // Collision with another path template with the same simple resource name; qualify this resource name.
-        Name qualifiedResourceName = namer.getQualifiedResourceIdentifier(namePattern);
+        Name qualifiedResourceName = DiscoGapicParser.getQualifiedResourceIdentifier(namePattern);
         methodToResourceName.put(method, qualifiedResourceName);
       } else {
         simpleResourceToFirstPatternMap.put(simpleResourceName, namePattern);
@@ -91,7 +90,7 @@ public class DiscoConfigTransformer {
         .license(generateLicense())
         .interfaces(generateInterfaces(model, methodToNamePattern.build(), methodToResourceNames))
         .resourceNameGeneration(
-            generateResourceNameGenerations(model.getDocument(), namer, methodToResourceNames))
+            generateResourceNameGenerations(model.getDocument(), methodToResourceNames))
         .build();
   }
 
@@ -169,14 +168,15 @@ public class DiscoConfigTransformer {
   }
 
   private List<ResourceNameGenerationView> generateResourceNameGenerations(
-      Document model, DiscoGapicNamer namer, Map<Method, Name> methodToResourceNameAndPatternMap) {
+      Document model, Map<Method, Name> methodToResourceNameAndPatternMap) {
     ImmutableList.Builder<ResourceNameGenerationView> resourceNames = ImmutableList.builder();
     for (Map.Entry<String, List<Method>> resource : model.resources().entrySet()) {
       for (Method method : resource.getValue()) {
         if (!Strings.isNullOrEmpty(method.path())) {
           ResourceNameGenerationView.Builder view = ResourceNameGenerationView.newBuilder();
-          view.messageName(namer.getRequestName(method).toUpperCamel());
-          String parameterName = namer.getResourceIdentifier(method.flatPath()).toLowerCamel();
+          view.messageName(DiscoGapicParser.getRequestName(method).toUpperCamel());
+          String parameterName =
+              DiscoGapicParser.getResourceIdentifier(method.flatPath()).toLowerCamel();
           String resourceName = methodToResourceNameAndPatternMap.get(method).toLowerCamel();
           Map<String, String> fieldEntityMap = new HashMap<>();
           fieldEntityMap.put(parameterName, resourceName);
