+++ src/main/java/com/google/api/codegen/discogapic/transformer/java/JavaDiscoGapicRequestToViewTransformer.java
@@ -29,6 +29,7 @@ import com.google.api.codegen.config.MethodModel;
 import com.google.api.codegen.config.PackageMetadataConfig;
 import com.google.api.codegen.discogapic.SchemaTransformationContext;
 import com.google.api.codegen.discogapic.transformer.DiscoGapicNamer;
+import com.google.api.codegen.discogapic.transformer.DiscoGapicParser;
 import com.google.api.codegen.discogapic.transformer.DocumentToViewTransformer;
 import com.google.api.codegen.discovery.Method;
 import com.google.api.codegen.discovery.Schema;
@@ -204,9 +205,7 @@ public class JavaDiscoGapicRequestToViewTransformer implements DocumentToViewTra
         context
             .getNamer()
             .privateFieldName(
-                context
-                    .getDiscoGapicNamer()
-                    .getRequestName(((DiscoveryMethodModel) method).getDiscoMethod()));
+                DiscoGapicParser.getRequestName(((DiscoveryMethodModel) method).getDiscoMethod()));
     String requestName =
         nameFormatter.privateFieldName(Name.anyCamel(symbolTable.getNewSymbol(requestClassId)));
     boolean hasRequiredProperties = false;
@@ -260,8 +259,7 @@ public class JavaDiscoGapicRequestToViewTransformer implements DocumentToViewTra
 
     StaticLangApiMessageView.Builder paramView = StaticLangApiMessageView.newBuilder();
     Method discoMethod = ((DiscoveryMethodModel) method).getDiscoMethod();
-    String resourceName =
-        context.getDiscoGapicNamer().getResourceIdentifier(discoMethod.path()).toLowerCamel();
+    String resourceName = DiscoGapicParser.getResourceIdentifier(discoMethod.path()).toLowerCamel();
     StringBuilder description =
         new StringBuilder(discoMethod.parameters().get(resourceName).description());
     description.append(String.format("\nIt must have the format `%s`. ", discoMethod.path()));
@@ -301,9 +299,7 @@ public class JavaDiscoGapicRequestToViewTransformer implements DocumentToViewTra
           schemaToParamView(
               context,
               requestBody,
-              context
-                  .getDiscoGapicNamer()
-                  .getSchemaNameAsParameter(requestBodyDef.dereference())
+              DiscoGapicParser.getSchemaNameAsParameter(requestBodyDef.dereference())
                   .toLowerCamel(),
               symbolTable,
               EscapeName.NO_ESCAPE_NAME));
@@ -368,7 +364,6 @@ public class JavaDiscoGapicRequestToViewTransformer implements DocumentToViewTra
     return new SchemaTypeTable(
         typeTable,
         typeNameConverter,
-        new DiscoGapicNamer(),
         new JavaSurfaceNamer(implicitPackageName, implicitPackageName));
   }
